Update to v0.2.2 changes in the Description
- Remove Profile Link to non-existent Profiles page on Blog Articles.
- Fix Accounts that are not Account Contacts for Hosting or Site Payments from showing up as Payment Accounts.
- Remove some Social Media Share To options on content pages.
- Fix images for Services/Events not displaying in the selector on Bookings Page.
- Adjust Administration Folder Viewing Permissions.
- Fix filtering of editor data for FAQ's, that was removing links from the data.
- Adjust Administration Menu and Header area to help with viewing on Mobile Devices.
- Adjust Coming Soon and Maintenance Pages, add Countdown Clock to Coming Soon page.
- Improve Administration Sidebar Hiding/Showing actions.
- Fix HTML Attribute order in Administration pages.
- Fix Menu not highlighting correct active page in header and footer menus.
- Add role values to header and footer navigation menus.
- Fix Inventory Stock Status not Showing item as Pre Order.
- Remove Tooltip trigger from Checkboxes in Administration but leave aria-label for Screen Readers.
- Fix Administration Notifications dropdown styling, and data retrieval.
- Adjust robots.txt file to exclude areas from being indexed by search engines.
- Fix sitemap.xml and sitemap adding links to pages that cause errors, or shouldn't get indexed.
- Fix Escaping Parser for JSON-LD.
- Stop Offline page showing, but rather routing to an error page, as it should only be used for offline circumstances, and not indexable.
- Fix content items from showing malformed content if content item title and URLSlug changes, now goes to a 404 Error Page.
- Adjust Error page to not fetch content that isn't available to non-logged in site visitors.
- Extract common functions into seperate files, such as breadcrumbs, cover image building, to overall reduce size, and allow easier future editing.
- Add HTMLPurifier to properly mitigate XSS code executing on HTML Output.
- Fix Gallery Items on Gallery and Content pages displaying wrong thumbnail.
- Make sure Gallery Item on Gallery and Content pages are responsive.
- Fix Bookings page showing non Service items.
- Shopping Cart Widget not displaying no image image correctly.
- Fix content Pagination not showing correct number of pages, or correct previous page link value.
- Fix Forum unset values throwing errors.
- Add missing role="menuitem" to login/administration menu links.
- Fix Page list in Administration not clearing views for all items.

+++ core/purge.php
@@ -7,7 +7,7 @@
  * @author     Dennis Suitters <dennis@diemen.design>
  * @copyright  2014-2019 Diemen Design
  * @license    http://opensource.org/licenses/MIT  MIT License
- * @version    0.1.9
+ * @version    0.2.2
  * @link       https://github.com/DiemenDesign/AuroraCMS
  * @notes      This PHP Script is designed to be executed using PHP 7+
  */
@@ -19,7 +19,7 @@ $tbl=isset($_POST['t'])?filter_input(INPUT_POST,'t',FILTER_SANITIZE_STRING):filt
 $col=isset($_POST['c'])?filter_input(INPUT_POST,'c',FILTER_SANITIZE_STRING):filter_input(INPUT_GET,'c',FILTER_SANITIZE_STRING);
 $uid=isset($_SESSION['uid'])?$_SESSION['uid']:0;
 $el='l_';
-if($id!=0&&$tbl!='logs'&&$tbl!='livechat'&&$tbl!='forumCategory'||$tbl!='forumTopics'){
+if($id!=0&&$tbl!='logs'&&$tbl!='livechat'&&$tbl!='forumCategory'&&$tbl!='forumTopics'&&$tbl!='pageviews'){
   $s=$db->prepare("SELECT * FROM `".$prefix.$tbl."` WHERE `id`=:id");
   $s->execute([':id'=>$id]);
   $r=$s->fetch(PDO::FETCH_ASSOC);
@@ -77,7 +77,7 @@ if($id!=0&&$tbl=='seo'){
   $q->execute([':id'=>$id]);
   $el='l_';
 }
-if($id!=0&&$id!='activity'){
+if($id!=0&&$id!='activity'&&$tbl!='pageviews'){
   $q=$db->prepare("DELETE FROM `".$prefix.$tbl."` WHERE `id`=:id");
   $q->execute([':id'=>$id]);
   if($tbl=='media')$el='media_items_';
